backend_django/notesproject/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==3.11.4
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.7.1
sqlparse==0.5.3
tzdata==2025.2
